+++ public/src/components/channelManagement/campaigns/TestDataButton.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Button, makeStyles } from '@material-ui/core';
 
 import { Test } from '../helpers/shared';
+import { formattedTimestamp } from '../helpers/utilities';
 import TestDataDialog from './TestDataDialog';
 import useOpenable from '../../../hooks/useOpenable';
 
@@ -19,12 +20,14 @@ const TestDataButton: React.FC<TestDataButtonProps> = ({ test }: TestDataButtonP
   const [isOpen, open, close] = useOpenable();
   const classes = useStyles();
 
+  const datetimeStamp = formattedTimestamp(Date());
+
   return (
     <>
       <Button className={classes.button} variant="contained" onClick={open}>
         View data
       </Button>
-      <TestDataDialog isOpen={isOpen} close={close} test={test} />
+      <TestDataDialog isOpen={isOpen} close={close} test={test} datetimeStamp={datetimeStamp} />
     </>
   );
 };
